+++ src/main/java/racingcar/dto/ProgressState.java
@@ -1,4 +1,4 @@
-package racingcar.domain.dto;
+package racingcar.dto;
 
 import java.util.List;
 import racingcar.domain.Car;
